requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==3.13.0
django-debug-toolbar==4.4.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
whitenoise==6.2.0
